Move random location sampling to transforms
